refactor(runtime): split scheduler eligibility files
