+++ pkg/services/factory_runtime/internal/services/orchestration/scheduler/enablement_runtime_test.go
@@ -48,6 +48,248 @@ func TestEnablementEvaluator_SameNameGuardDeduplicatesRegisteredParentCandidates
 	}
 }
 
+func TestEnablementEvaluator_SameNameGuardUsesOrderedCurrentParentChild(t *testing.T) {
+	eval := NewEnablementEvaluator(nil, testNow, nil)
+	n := sameNameGuardNet()
+	historical := &factorytoken.Token{
+		ID:      "a-token-history",
+		PlaceID: "task:ready",
+		Color: factorytoken.Color{
+			Name:       "resume-project",
+			WorkID:     "cycle-44",
+			WorkTypeID: "project-cycle",
+			ParentID:   "project-work",
+		},
+	}
+	current := *historical
+	current.ID = "z-token-current"
+	current.Color.WorkID = "cycle-93"
+	marking := makeTestSnapshot(map[string]*factorytoken.Token{
+		"project-token": {
+			ID:      "project-token",
+			PlaceID: "plan:ready",
+			Color: factorytoken.Color{
+				Name:       "resume-project",
+				WorkID:     "project-work",
+				WorkTypeID: "project",
+			},
+		},
+		historical.ID: historical,
+		current.ID:    &current,
+	})
+	marking.ParentChildRegistrations = petri.ParentChildRegistrationProjection{
+		"project-work": {Children: []factorytoken.Token{*historical, current}, Complete: true},
+	}
+
+	enabled := eval.FindEnabledTransitions(context.Background(), n, &marking)
+	if len(enabled) != 1 {
+		t.Fatalf("enabled transitions = %d, want 1", len(enabled))
+	}
+	if got := tokenIDs(enabled[0].Bindings["task"]); strings.Join(got, ",") != current.ID {
+		t.Fatalf("current cycle binding = %v, want [%s] despite token ID ordering", got, current.ID)
+	}
+}
+
+func TestEnablementEvaluator_SameNameGuardOnParentUsesOrderedCurrentChild(t *testing.T) {
+	eval := NewEnablementEvaluator(nil, testNow, nil)
+	n := &state.Net{
+		Places: map[string]*petri.Place{
+			"project:waiting":       {ID: "project:waiting"},
+			"project-cycle:blocked": {ID: "project-cycle:blocked"},
+		},
+		Transitions: map[string]*petri.Transition{
+			"block-project": {
+				ID:   "block-project",
+				Name: "block-project",
+				InputArcs: []petri.Arc{
+					{
+						ID:          "project-in",
+						Name:        "project",
+						PlaceID:     "project:waiting",
+						Direction:   petri.ArcInput,
+						Cardinality: petri.ArcCardinality{Mode: petri.CardinalityOne},
+						Guard:       &petri.SameNameGuard{MatchBinding: "project-cycle"},
+					},
+					{
+						ID:          "cycle-in",
+						Name:        "project-cycle",
+						PlaceID:     "project-cycle:blocked",
+						Direction:   petri.ArcInput,
+						Cardinality: petri.ArcCardinality{Mode: petri.CardinalityOne},
+					},
+				},
+			},
+		},
+	}
+	historical := &factorytoken.Token{
+		ID:      "a-token-history",
+		PlaceID: "project-cycle:blocked",
+		Color: factorytoken.Color{
+			Name:       "resume-project",
+			WorkID:     "cycle-44",
+			WorkTypeID: "project-cycle",
+			ParentID:   "project-work",
+		},
+	}
+	current := *historical
+	current.ID = "z-token-current"
+	current.Color.WorkID = "cycle-93"
+	parent := &factorytoken.Token{
+		ID:      "project-token",
+		PlaceID: "project:waiting",
+		Color: factorytoken.Color{
+			Name:       "resume-project",
+			WorkID:     "project-work",
+			WorkTypeID: "project",
+		},
+	}
+	marking := makeTestSnapshot(map[string]*factorytoken.Token{
+		parent.ID:     parent,
+		historical.ID: historical,
+		current.ID:    &current,
+	})
+	marking.ParentChildRegistrations = petri.ParentChildRegistrationProjection{
+		"project-work": {Children: []factorytoken.Token{*historical, current}, Complete: true},
+	}
+
+	enabled := eval.FindEnabledTransitions(context.Background(), n, &marking)
+	if len(enabled) != 1 {
+		t.Fatalf("enabled transitions = %d, want 1", len(enabled))
+	}
+	if got := tokenIDs(enabled[0].Bindings["project-cycle"]); strings.Join(got, ",") != current.ID {
+		t.Fatalf("current cycle binding = %v, want [%s] despite parent guard orientation", got, current.ID)
+	}
+}
+
+func TestEnablementEvaluator_SameNameGuardDoesNotFallbackToHistoricalChild(t *testing.T) {
+	eval := NewEnablementEvaluator(nil, testNow, nil)
+	n := &state.Net{
+		Places: map[string]*petri.Place{
+			"project:waiting":       {ID: "project:waiting"},
+			"project-cycle:blocked": {ID: "project-cycle:blocked"},
+			"project-cycle:init":    {ID: "project-cycle:init"},
+		},
+		Transitions: map[string]*petri.Transition{
+			"block-project": {
+				ID:   "block-project",
+				Name: "block-project",
+				InputArcs: []petri.Arc{
+					{
+						ID:          "project-in",
+						Name:        "project",
+						PlaceID:     "project:waiting",
+						Direction:   petri.ArcInput,
+						Cardinality: petri.ArcCardinality{Mode: petri.CardinalityOne},
+						Guard:       &petri.SameNameGuard{MatchBinding: "project-cycle"},
+					},
+					{
+						ID:          "cycle-in",
+						Name:        "project-cycle",
+						PlaceID:     "project-cycle:blocked",
+						Direction:   petri.ArcInput,
+						Cardinality: petri.ArcCardinality{Mode: petri.CardinalityOne},
+					},
+				},
+			},
+		},
+	}
+	historical := &factorytoken.Token{
+		ID:      "a-token-history",
+		PlaceID: "project-cycle:blocked",
+		Color: factorytoken.Color{
+			Name:       "resume-project",
+			WorkID:     "cycle-44",
+			WorkTypeID: "project-cycle",
+			ParentID:   "project-work",
+		},
+	}
+	current := *historical
+	current.ID = "z-token-current"
+	current.PlaceID = "project-cycle:init"
+	current.Color.WorkID = "cycle-93"
+	parent := &factorytoken.Token{
+		ID:      "project-token",
+		PlaceID: "project:waiting",
+		Color: factorytoken.Color{
+			Name:       "resume-project",
+			WorkID:     "project-work",
+			WorkTypeID: "project",
+		},
+	}
+	marking := makeTestSnapshot(map[string]*factorytoken.Token{
+		parent.ID:     parent,
+		historical.ID: historical,
+		current.ID:    &current,
+	})
+	marking.ParentChildRegistrations = petri.ParentChildRegistrationProjection{
+		"project-work": {Children: []factorytoken.Token{*historical, current}, Complete: true},
+	}
+
+	if enabled := eval.FindEnabledTransitions(context.Background(), n, &marking); len(enabled) != 0 {
+		t.Fatalf("historical child enabled block-project while current child was in init: %#v", enabled)
+	}
+}
+
+func TestEnablementEvaluator_SameNameParentGuardFailsClosedForInvalidRegistration(t *testing.T) {
+	for _, test := range []struct {
+		name          string
+		complete      bool
+		contradictory bool
+	}{
+		{name: "incomplete registration"},
+		{name: "contradictory registration", complete: true, contradictory: true},
+	} {
+		t.Run(test.name, func(t *testing.T) {
+			eval := NewEnablementEvaluator(nil, testNow, nil)
+			n := sameNameGuardNet()
+			transition := n.Transitions["match-items"]
+			transition.InputArcs[0].Guard = &petri.SameNameGuard{MatchBinding: "task"}
+			transition.InputArcs[1].Guard = nil
+
+			parent := &factorytoken.Token{
+				ID:      "project-token",
+				PlaceID: "plan:ready",
+				Color: factorytoken.Color{
+					Name:   "resume-project",
+					WorkID: "project-work",
+				},
+			}
+			historical := &factorytoken.Token{
+				ID:      "a-token-history",
+				PlaceID: "task:ready",
+				Color: factorytoken.Color{
+					Name:     "resume-project",
+					WorkID:   "cycle-44",
+					ParentID: "project-work",
+				},
+			}
+			current := *historical
+			current.ID = "z-token-current"
+			current.Color.WorkID = "cycle-93"
+
+			registeredCurrent := current
+			if test.contradictory {
+				registeredCurrent.Color.ParentID = "other-parent"
+			}
+			marking := makeTestSnapshot(map[string]*factorytoken.Token{
+				parent.ID:     parent,
+				historical.ID: historical,
+				current.ID:    &current,
+			})
+			marking.ParentChildRegistrations = petri.ParentChildRegistrationProjection{
+				"project-work": {
+					Children: []factorytoken.Token{*historical, registeredCurrent},
+					Complete: test.complete,
+				},
+			}
+
+			if enabled := eval.FindEnabledTransitions(context.Background(), n, &marking); len(enabled) != 0 {
+				t.Fatalf("invalid registration enabled historical same-name transition: %#v", enabled)
+			}
+		})
+	}
+}
+
 func TestSameNameMatchBinding_HandlesNestedAndMalformedGuards(t *testing.T) {
 	valid := &petri.SameNameGuard{MatchBinding: "parent"}
 	nested := &petri.AllGuard{Guards: []petri.Guard{&petri.DependencyGuard{}, valid}}
